Make the Linux build warning-free: 1403 warnings hid 11 real bugs
The build emitted 1403 unique warning sites and carried two -Wno- flags that hid 21
more. 98% of that was one category - -Wwrite-strings, 1370 sites, string literals
passed to char*/wchar_t* - which meant the genuinely interesting diagnostics were
unreadable in the noise. Both targets now build clean with GCC's default diagnostics
and no suppressions, verified two ways: a from-scratch build of all 1157 translation
units (zero compiler and zero linker warnings), and independently by replaying every
compile_commands.json entry with -fsyntax-only and the -Wno- flags stripped.

The 1370 write-strings sites collapsed to four declarations plus seven static name
tables, and no call site needed touching: CompoundTag's ~26 NBT accessors (~671
sites), the four PIX* instrumentation stubs (~200), Tag::getTagName (13), and the
ColourTable/Textures/GameRuleManager/SoundEngine/DLC*/SkinSelect tables (481). All
fixed with const, which is source-compatible with MSVC - the two shared headers
involved (CompoundTag.h, x64headers/extraX64.h) read as corrections, not shims.

The point of doing it is what the noise was hiding. Two matter:

LONG_MAX is not RECT's LONG max on LP64. UIController.cpp seeded the custom-render
clear rect's min/max accumulator with LONG_MAX/LONG_MIN, which are `long`'s limits.
They coincide only on Windows, where LONG *is* a 32-bit long; here they truncate to
-1/0, so setupCustomDrawMatrices started out already inverted and the depth clear in
endCustomDrawGameState got a garbage rect for every Iggy custom-draw region. Broken on
Orbis too. Now (std::numeric_limits<LONG>::max)() - parenthesised, because windows.h
makes min/max macros.

A `new byte[]` buffer freed with scalar `delete` through a LPVOID, on the
load-a-save/tutorial path (MinecraftServer.cpp:264). That is a fourth instance of
KNOWN_BUGS.md's heap-corruption class, and -Wdelete-incomplete finds it for free.
Three more delete-through-void* sites (GameNetworkManager, two KickPlayerReturned
callbacks) plus MeleeAttackGoal deleting an incomplete `Path`, so ~Path() never ran on
either site and the whole node array leaked.

The rest: `title == L"Controls"` was an assignment written as a comparison, so the
screen title was never set; both local-socket close() methods called
s_hostQueue[i].empty() for a discarded result where clearing was meant, so a closed
stream kept everything still buffered on it; GameMode::useItem had no `return` at all
(shadowed in practice - Minecraft::gameMode is a MultiPlayerGameMode*, a separate
class despite the name - so this is UB removed, not behaviour changed); six int-to-
pointer casts that lose bits on LP64 now widen through intptr_t; and NULL used as an
integer, as an int[][7] element, as an IggyLibrary (a typedef'd S32) and as
SetFilePointer's dwMoveMethod.

Removing -Wno-template-body meant fixing what it described: ListTag<T>::print called
Tag::print(prefix, out) when the only overloads were print(ostream) and print(char*,
wostream), and all three took the stream by value, which cannot compile if called.
Dead debug dumpers that survived because MSVC never instantiates an unused member
template. Made consistently wide rather than deleted. Removing
-Wno-non-c-typedef-for-linkage meant rewriting 19 `typedef struct { ... } Name;`
hash/eq functors as plain structs - mechanical.

One deliberate behaviour change: Villager::addOffers used std::random_shuffle, which
is deprecated and drew from the global rand(), so trade order already differed between
platforms and between runs on one seed. Now Fisher-Yates over the villager's own
Random, deterministic like every other random decision the mob makes.
std::binary_function bases on the two chunk sorters are dropped; nothing read the
typedefs.

One suppression is deliberately kept: -Wno-unused-function/-variable on
Linux/Iggy/gdraw_sdl.c, which compiles the vendor .inl files as-is. Measured to hide
nothing at this project's warning level (both need -Wall), so it costs no visibility
today and guards whoever turns -Wall on later.

Verified at runtime as well as at build time: the client boots, generates, renders and
exits cleanly twice, with the HUD and the Iggy custom-draw tooltip - the clear-rect
path - correct. Not playtested specifically: a villager trade, and item-use-in-hand.
KNOWN_BUGS.md records the sweep method, including that hunks must be deduped by
realpath:line:col or header warnings multiply by TU count.

+++ Minecraft.Client/Common/UI/UIScene_HowToPlayMenu.cpp
@@ -200,6 +200,8 @@ void UIScene_HowToPlayMenu::handlePress(F64 controlId, F64 childId)
 
 		unsigned int uiInitData;
 		uiInitData = ( ( 1 <<  31 )  | ( m_uiHTPSceneA[(int)childId]  << 16 ) | ( short )( m_iPad ) );
-		ui.NavigateToScene(m_iPad, eUIScene_HowToPlay, ( void* )( uiInitData ) );
+		// uiInitData is a packed word smuggled through the void* init-data slot,
+		// not a real pointer - widen it explicitly.
+		ui.NavigateToScene(m_iPad, eUIScene_HowToPlay, ( void* )( intptr_t )( uiInitData ) );
 	}
 }
