+++ Minecraft.World/TheEndPortal.cpp
@@ -18,7 +18,7 @@ bool TheEndPortal::allowAnywhere()
 
 void TheEndPortal::allowAnywhere(bool set)
 {
-	TlsSetValue(tlsIdx,(LPVOID)(set?1:0));
+	TlsSetValue(tlsIdx,(LPVOID)(intptr_t)(set?1:0));	// widen through intptr_t: int -> pointer
 }
 
 TheEndPortal::TheEndPortal(int id, Material *material) : EntityTile(id, material, isSolidRender())
